requirements.txt
django==3.1.2
django-phonenumber-field[phonenumberslite]
gunicorn
django_heroku
pillow
whitenoise